crit: add core-dump feature, v3
crit core-dump is to be used to produce valid elf core dump
from criu images. It is already able to produce core dump that
is readable and can be used by objdump, readelf and gdb.
Still, it is not perfect and lacks some vma contents(vsyscall
and vvar) and some core notes are not fullfilled.

It is also worth noting that for now only x86_64 architecture
is supported.

v2, fixed memory pages search
v3, get cmdline from pages, improve memory handling, add more
    helpers and add more comments